🚚 rename: 취소확인 및 정보 모달을 modal 폴더로 이동 #58
